+++ src/components/VibeFormDatepicker.vue
@@ -59,14 +59,15 @@ const {
   ariaDescribedBy,
   shouldRenderLabel,
   shouldRenderFeedback,
-  shouldRenderHelp
+  shouldRenderHelp,
+  validationClass,
+  ariaInvalid
 } = useFormField('datepicker', props)
 
 const inputClass = computed(() => {
   const classes = ['form-control']
   if (props.size) classes.push(`form-control-${props.size}`)
-  if (props.validationState === 'valid') classes.push('is-valid')
-  if (props.validationState === 'invalid') classes.push('is-invalid')
+  if (validationClass.value) classes.push(validationClass.value)
   return classes.join(' ')
 })
 
@@ -109,7 +110,7 @@ const handleFocus = (event: FocusEvent) => {
       :required="required"
       :min="min"
       :max="max"
-      :aria-invalid="validationState === 'invalid'"
+      :aria-invalid="ariaInvalid"
       :aria-describedby="ariaDescribedBy"
       @input="handleInput"
       @change="handleChange"